requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
idna==3.7
pillow==10.3.0
PyJWT==2.8.0
python-decouple==3.8
redis==5.0.4
requests==2.31.0
six==1.16.0
spotipy==2.23.0
sqlparse==0.5.0
urllib3==2.2.1
